requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic
gradio
pydantic-ai
logfire
python-dotenv
beautifulsoup4
requests
tavily-python
webdriver-manager
selenium